+++ libft/ft_strlen.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-int		ft_strlen(char *s)
+size_t		ft_strlen(char *s)
 {
-	int		i;
+	size_t		i;
 
 	i = 0;
 	while (s[i] != '\0')
